Added Library Page
Added the Library page to the dashboard utilizing an array to store book objects from the user inputs. The user can
add a new book via a form by using the Add Book button - use of modal. When a book has been read, the left of the card display turns green, and the unread remains yellow. Existing read/unread books can be edited or deleted.